requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
pycryptodomex==3.20.0
pymongo==4.8.0
pyTelegramBotAPI==4.21.0
pytube2==15.0.6
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
